Fix bug and re-structure
